requirements.txt
# Web Framework
Flask==3.0.0
flask-cors==4.0.0

# Database
supabase==2.3.0

# Machine Learning & Computer Vision
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3

# Data Processing
pandas==2.1.3
openpyxl==3.1.5

# Environment Variables
python-dotenv==1.0.0
